Use in-memory distance matrix by default

+++ bin_x/cli/clustering.py
@@ -9,7 +9,10 @@ import pandas as pd
 from numpy.lib.format import open_memmap  # noqa
 
 from bin_x.core.clustering.algorithm import fit_cluster
-from bin_x.core.clustering.distance_matrix import create_distance_matrix
+from bin_x.core.clustering.distance_matrix import (
+    create_distance_matrix,
+    create_in_mem_distance_matrix,
+)
 from bin_x.core.clustering.dump_bins import dump_bins
 
 
@@ -21,6 +24,7 @@ def perform_clustering(
     max_iterations: int = 10,
     metric: str = "convex",
     qp_solver: str = "quadprog",
+    in_mem_dist_matrix: bool = True,
     distance_matrix_cache: Optional[Path] = None,
 ) -> Path:
     """
@@ -33,7 +37,9 @@ def perform_clustering(
     :param max_iterations: Number of maximum iterations to perform.
     :param metric: Polytope distance matrix (convex/affine)
     :param qp_solver: Quadratic programming problem solver. (quadprog/cvxopt)
+    :param in_mem_dist_matrix: Whether to use in memory distance matrix.
     :param distance_matrix_cache: Previously computed distance matrix.
+                                    in_mem_dist_matrix must be false if this is provided.
     :return: Path of the binning result dataset.
     """
     dist_bin_csv = operating_dir / "binning-assignment.csv"
@@ -52,12 +58,17 @@ def perform_clustering(
 
     # 02. Create a distance matrix
     distance_matrix_filename = distance_matrix_cache
-    if distance_matrix_filename is None:
-        click.secho(f">> Creating a distance matrix of {num_samples}x{num_samples} shape...", fg="green", bold=True)
-        distance_matrix_filename = create_distance_matrix(samples, operating_dir)
+    if in_mem_dist_matrix:
+        assert distance_matrix_filename is None, "Distance matrix cache cannot be used in in-memory mode"
+        click.secho(f">> Creating the distance matrix({num_samples}x{num_samples}) in-memory...", fg="green", bold=True)
+        distance_matrix = create_in_mem_distance_matrix(samples)
     else:
-        click.secho(f">> Reusing distance matrix at {distance_matrix_filename}...", fg="green", bold=True)
-    distance_matrix = open_memmap(filename=distance_matrix_filename, mode="r", shape=(num_samples, num_samples))
+        if distance_matrix_filename is None:
+            click.secho(f">> Creating a distance matrix of {num_samples}x{num_samples} shape...", fg="green", bold=True)
+            distance_matrix_filename = create_distance_matrix(samples, operating_dir)
+        else:
+            click.secho(f">> Reusing distance matrix at {distance_matrix_filename}...", fg="green", bold=True)
+        distance_matrix = open_memmap(filename=distance_matrix_filename, mode="r", shape=(num_samples, num_samples))
 
     # 03. Perform binning using specified solver and metric
     click.secho(f">> Performing binning using {qp_solver} solver", fg="green", bold=True)
@@ -73,7 +84,7 @@ def perform_clustering(
     )
 
     # Delete the distance matrix file (dont delete if using a cache)
-    if distance_matrix_cache is None:
+    if not in_mem_dist_matrix and distance_matrix_cache is None:
         os.remove(distance_matrix_filename)
     if np.any(convex_labels < 0):
         raise ValueError("There were some un-clustered points left... Aborting.")
@@ -124,5 +135,6 @@ def run_perform_clustering(
         max_iterations=int(parameters["AlgoMaxIterations"]),
         metric=parameters["AlgoDistanceMetric"],
         qp_solver=parameters["AlgoQpSolver"],
+        in_mem_dist_matrix=parameters.getboolean("InMemDistMatrix"),
         distance_matrix_cache=distance_matrix_cache,
     )
